Documentation, Import fixes
All of the import statements inside of core, counter and js have been updated so that they use the std module importer.  This makes it easier to work with spork as a separate module, un-hooked from the standard one.

Also, I have started off with some documentation.  The org file is exported into the html version.  If needed I can also export to a text file